pyproject.toml
[project]
name = "causal-armor-langgraph-demo"
version = "0.1.0"
description = "Demo: LangGraph agent with and without CausalArmor protection"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.3",
    "pypdf>=4.0",
    "causal-armor @ git+https://github.com/prashantkul/causal-armor.git",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "fpdf2>=2.7"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/demo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
